ranking groups into a different view
